move: Loader components/ui로 이동 (#10)
