+++ src/pages/form/my/response/Answer.tsx
@@ -1,4 +1,4 @@
-import { Loader } from '@components/Loader';
+import { Loader } from '@/components/ui/Loader';
 import { useGetMyAnswer } from '@query-hooks/form';
 
 export const Answer = ({ questionId }: { questionId: number }) => {
