+++ src/pages/form/upload/_FormUploadPage.tsx
@@ -3,7 +3,7 @@ import { useEffect, useReducer, useState } from 'react';
 import { Box } from '@/components/ui/Box';
 import { Period } from '@src/types/period-picker/period';
 import { FormUpload, Question } from '@api/form-upload/types';
-import { SubmitButton } from '@components/SubmitButton';
+import { SubmitButton } from '@components/ui/Button';
 import { useTopBarStore } from '@stores/topBar-stores';
 import { useModal } from '@/hooks/modal/useModal';
 import { useCreateForm } from '@query-hooks/form-upload';
